restore createwallet command
