+++ cli/walletloader/walletloader.go
@@ -1,4 +1,4 @@
-package runner
+package walletloader
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 )
 
 // createWallet creates a new wallet if one doesn't already exist using the WalletMiddleware provided
-func createWallet(ctx context.Context, walletMiddleware app.WalletMiddleware) (err error) {
+func CreateWallet(ctx context.Context, walletMiddleware app.WalletMiddleware) (err error) {
 	// first check if wallet already exists
 	walletExists, err := walletMiddleware.WalletExists()
 	if err != nil {
@@ -98,25 +98,7 @@ func createWallet(ctx context.Context, walletMiddleware app.WalletMiddleware) (e
 	return SyncBlockChain(ctx, walletMiddleware)
 }
 
-// displayWalletSeed prints the generated seed for a new wallet
-func displayWalletSeed(seed string) {
-	fmt.Println("Your wallet generation seed is:")
-	fmt.Println("-------------------------------")
-	seedWords := strings.Split(seed, " ")
-	for i, word := range seedWords {
-		fmt.Printf("%s ", word)
-
-		if (i+1)%6 == 0 {
-			fmt.Printf("\n")
-		}
-	}
-	fmt.Println("\n-------------------------------")
-	fmt.Println("IMPORTANT: Keep the seed in a safe place as you will NOT be able to restore your wallet without it.")
-	fmt.Println("Please keep in mind that anyone who has access to the seed can also restore your wallet thereby " +
-		"giving them access to all your funds, so it is imperative that you keep it in a secure location.")
-}
-
-// openWallet is called whenever an action to be executed requires wallet to be loaded
+// OpenWallet is called whenever an action to be executed requires wallet to be loaded
 // notifies the program to exit if wallet doesn't exist or some other error occurs by returning a non-nil error
 //
 // this method may stall until previous godcr instances are closed (especially in cases of multiple mobilewallet instances)
@@ -167,31 +149,6 @@ func OpenWallet(ctx context.Context, walletMiddleware app.WalletMiddleware) (wal
 		err = ctx.Err()
 		return
 	}
-}
-
-func attemptToCreateWallet(ctx context.Context, walletMiddleware app.WalletMiddleware) error {
-	createWalletPrompt := "No wallet found. Would you like to create one now? (y/N)"
-	validateUserResponse := func(userResponse string) error {
-		userResponse = strings.TrimSpace(userResponse)
-		userResponse = strings.Trim(userResponse, `"`)
-		if userResponse == "" || strings.EqualFold("y", userResponse) || strings.EqualFold("N", userResponse) {
-			return nil
-		} else {
-			return fmt.Errorf("invalid option, try again")
-		}
-	}
-	userResponse, err := terminalprompt.RequestInput(createWalletPrompt, validateUserResponse)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading your response: %s", err.Error())
-		return err
-	}
-
-	if userResponse == "" || strings.EqualFold("N", userResponse) {
-		fmt.Println("Maybe later. Bye.")
-		return nil
-	}
-
-	return createWallet(ctx, walletMiddleware)
 }
 
 // syncBlockChain uses the WalletMiddleware provided to download block updates
